Cleaned up API a bit, hid some things
