+++ lib/src/application/application_server.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:isolate';
-import 'dart:mirrors';
 
 import 'package:logging/logging.dart';
 
@@ -9,20 +7,42 @@ import '../http/request.dart';
 import '../http/request_sink.dart';
 import 'application.dart';
 import 'application_configuration.dart';
-import 'isolate_supervisor.dart';
 
-/// Used internally.
+/// Represents a [RequestSink] manager being used by an [Application].
+///
+/// An Aqueduct application creates instances of this type to pair an HTTP server and an
+/// instance of an application-specific [RequestSink].
 class ApplicationServer {
+
+  /// The configuration this instance used to start its [sink].
   ApplicationConfiguration configuration;
+
+  /// The underlying [HttpServer].
   HttpServer server;
+
+  /// The instance of [RequestSink] serving requests.
   RequestSink sink;
+
+  /// The unique identifier of this instance.
+  ///
+  /// Each instance has its own identifier, a numeric value starting at 1, to identify it
+  /// among other instances.
   int identifier;
+
+  /// The logger of this instance
   Logger get logger => new Logger("aqueduct");
 
+  /// Creates an instance of this type.
+  ///
+  /// You should not need to invoke this method directly.
   ApplicationServer(this.sink, this.configuration, this.identifier) {
     sink.server = this;
   }
 
+  /// Starts this instance, allowing it to receive HTTP requests.
+  ///
+  /// Do not invoke this method directly, [Application] instances are responsible
+  /// for calling this method.
   Future start({bool shareHttpServer: false}) async {
     try {
       sink.setupRouter(sink.router);
@@ -49,6 +69,13 @@ class ApplicationServer {
     }
   }
 
+  /// Invoked when this server becomes ready receive requests.
+  ///
+  /// This method will invoke [RequestSink.open] and await for it to finish.
+  /// Once [RequestSink.open] completes, the underlying [server]'s HTTP requests
+  /// will be sent to this instance's [sink].
+  ///
+  /// [RequestSink.didOpen] is invoked after this opening has completed.
   Future didOpen() async {
     logger.info("Server aqueduct/$identifier started.");
 
@@ -64,52 +91,4 @@ class ApplicationServer {
 
     sink.didOpen();
   }
-}
-
-/// Used internally.
-class ApplicationIsolateServer extends ApplicationServer {
-  SendPort supervisingApplicationPort;
-  ReceivePort supervisingReceivePort;
-
-  ApplicationIsolateServer(
-      RequestSink sink,
-      ApplicationConfiguration configuration,
-      int identifier,
-      this.supervisingApplicationPort)
-      : super(sink, configuration, identifier) {
-    sink.server = this;
-    supervisingReceivePort = new ReceivePort();
-    supervisingReceivePort.listen(listener);
-  }
-
-  @override
-  Future didOpen() async {
-    await super.didOpen();
-
-    supervisingApplicationPort.send(supervisingReceivePort.sendPort);
-  }
-
-  void listener(dynamic message) {
-    if (message == ApplicationIsolateSupervisor.MessageStop) {
-      server.close(force: true).then((s) {
-        supervisingApplicationPort
-            .send(ApplicationIsolateSupervisor.MessageStop);
-      });
-    }
-  }
-}
-
-/// This method is used internally.
-void isolateServerEntryPoint(ApplicationInitialServerMessage params) {
-  var sinkSourceLibraryMirror =
-      currentMirrorSystem().libraries[params.streamLibraryURI];
-  var sinkTypeMirror = sinkSourceLibraryMirror.declarations[
-      new Symbol(params.streamTypeName)] as ClassMirror;
-
-  var app = sinkTypeMirror.newInstance(
-      new Symbol(""), [params.configuration.configurationOptions]).reflectee;
-
-  var server = new ApplicationIsolateServer(
-      app, params.configuration, params.identifier, params.parentMessagePort);
-  server.start(shareHttpServer: true);
 }
