+++ lib/src/db/managed/managed.dart
@@ -1,7 +1,6 @@
 export 'attributes.dart';
 export 'context.dart';
 export 'data_model.dart';
-export 'data_model_builder.dart';
 export 'entity.dart';
 export 'object.dart';
 export 'property_description.dart';
