+++ test/base/pattern_test.dart
@@ -2,6 +2,8 @@ import "package:test/test.dart";
 import "dart:core";
 import 'package:aqueduct/aqueduct.dart';
 
+import '../../lib/src/http/route_node.dart';
+
 void main() {
   group("Pattern splitting", () {
     test("No optionals, no expressions", () {
